+++ 2024 Projects/C++ SFML/SFML_Projects/Platformer/TileData.cpp
@@ -1,6 +1,6 @@
 #include "TileData.h"
 
-TileData::TileData(sf::Texture* texture, sf::Vector2f position, int scaler, int tileID, sf::Vector2f arrayPos) :
+TileData::TileData(sf::Texture* texture, sf::Vector2f position, float scaler, int tileID, sf::Vector2f arrayPos) :
 	texture(texture), position(position), tileID(tileID), arrayPos(arrayPos)
 {
 	sprite.setTexture(*texture);
